fix: expose private API on private port (#9)
Because

- Private gRPC API was being exposed on the public port

This commit

- Uses the private port to serve the Private Artifact API
- Removes the exposure of the private port to the docker host.

# Notes

Other services listen and serve both the public and the private API
through the grpc-gateway. In the case of this service this isn't needed
for the private API as it doesn't expose any HTTP endpoints (other
services don't need to, either, but it's already exposed). Therefore,
the "listen and serve" strategy is a bit different. The gRPC config
should already contain the TLS information, if present, so we don't need
an if/else block for the private API.