fix(artifact): fix retry file process (#97)
Because

1. when all processing workers are busy, the dispatch worker continues
to send tasks to the queue. These tasks are not held by any workers, so
the dispatcher may keep sending the same tasks to the queue.

2. when chunk text is empty, it still insert into db.

This commit

1. **Add a check at the beginning of each process step** to verify the
file’s processing state in the database. If the state does not matches
the current process state, discard the file process task.

2. **Set buffer of worker queue to zero** to prevent out of date file
status being queued

3. **Skip the chunk's text if it text is empty**