feat(file): allow duplicate filenames (#218)
Because

- Filename shouldn't be unique in a catalog.

This commit

- Removes the dependencies on the filename as a unique identifier for a
file within a catalog and allows multiple files with the same filename.
- Endpoints that identified a file by (catalog, filename) now take a
file UID param. The filename params will be cleaned up after all the
clients adopt the file UID new params.
- On Milvus, we store the file UID as metadata and use it to query
similar embeddings. Legacy catalogs will still use the filename metadata
as the schema can't be updated.
- The unique index for (filename,catalog) in the database is dropped.
The database schema version is removed from the config as it's tied to
the code, not the external config.