chore(vector): keep collection name in the milvus package
The collection name is an implementation detail and, most importantly,
it is determined by the knowledge base UID. This commit updates the
vector database interface to take this value instead of computing the
collection name from it. The embedding.collection column is dropped in
the database as it's redundant with the file and knowledge base
information.